can_receive
更改灯板can接受协议

+++ Moudle/LightBand.h
@@ -7,7 +7,7 @@
 struct	lightband
 {
 	uint8_t   id;
-	uint8_t 	site;
+	float 		site;
 	uint8_t 	distance;
 	uint8_t 	direction;
